[OPR-278]: Starts adding Stress Tests
This is a new test bundle which gets tested as part of the OS X test suite.
